backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
openai>=1.12.0
anyio>=4.5.0
groq
mcp>=1.2.0
python-dotenv
greenlet
PyJWT==2.8.0